+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - W6401 `deprecated-pytest-yield-fixture`: add warning for [yield_fixture functions](https://docs.pytest.org/en/latest/yieldfixture.html)
 
+### Fixed
+- Fix incorrect path separator for Windows (#1)
+
 ## [0.2.0] - 2020-05-25
 ### Added
 - Suppressing FP `no-member` from [using workaround of accessing cls in setup fixture](https://github.com/pytest-dev/pytest/issues/3778#issuecomment-411899446)